better code structure, silk outline is now fully outside body dimensions
