+++ scripts/phoenix_contact/mc_params.py
@@ -328,6 +328,7 @@ class globalParams():
     mount_hole_layers = ['*.Cu', '*.Mask']
 
     silkGab = pin_Sx/2.0+0.05
+    silk_body_offset = 0.08
 
 
 #lock_cutout=
@@ -346,7 +347,7 @@ def dimensions(params):
         ,mount_hole_left = [-params.mount_hole_to_pin,mount_hole_y]
         ,mount_hole_right = [(params.num_pins-1)*params.pin_pitch+params.mount_hole_to_pin,mount_hole_y]
         ,flange_width = 9.2 if params.angled else 6.0
-        ,plug_front = width + params.back_to_pin -0.8
+        ,plug_front = width + params.back_to_pin -0.7
         ,plug_back = params.back_to_pin+0.6+0.25
     )
